starter_files/requirements.txt
flask
pillow